[libstagefright] fix -Wimplicit-int-float-conversion
IEEE754 single precision cannot precisely represent
INT32_MAX/UINT_MAX/RAND_MAX.

IEEE754 double precision cannot precisely represent INT64_MAX.

As such, implicit conversions from operations with floating point and
integral operands may result in a loss of precision.

Accept the imprecision with by making the existing implicit casts
explicit.

Bug: 139945549
Test: mm
Merged-in: Ic66b3d3bbac9e644e27ad8233018f24dfce3209f

+++ media/libstagefright/Utils.cpp
@@ -1806,7 +1806,7 @@ void convertMessageToMetaData(const sp<AMessage> &msg, sp<MetaData> &meta) {
     if (msg->findInt32("frame-rate", &fps) && fps > 0) {
         meta->setInt32(kKeyFrameRate, fps);
     } else if (msg->findFloat("frame-rate", &fpsFloat)
-            && fpsFloat >= 1 && fpsFloat <= INT32_MAX) {
+            && fpsFloat >= 1 && static_cast<int32_t>(fpsFloat) <= INT32_MAX) {
         // truncate values to distinguish between e.g. 24 vs 23.976 fps
         meta->setInt32(kKeyFrameRate, (int32_t)fpsFloat);
     }
